my Project
FOREST CONSERVATION
forests promote resilience since they can offer safety nets for livelihoods and are usually consumed more frequently when food is scarce. When they are sold, forest and tree products can help the world's food system become more nutrient-sensitive. minimizing water and air pollution. preserving open spaces and natural areas. conserving biodiversity, endangered species, and habitats for fish and animals. managing and defending wetlands and watersheds. maintaining leisure facilities and beautiful environments.

MY NAME IS Gerald-Kaparo Am a statistician by profession. Am currently learning to be a full stack developer at Moringa school in partnership with Flatiron.

The objectives of this policy document are to promote sustainable exploitation of forest resources for based on environmental protection and preservation and for the benefits of current and future generations of Kenyans. They also include respecting socio-cultural values and ensuring access to justice, gender equity, and inclusiveness. Finally, they aim to sustainably conserve water catchments areas by facilitating huma migration.

https://gerald-kaparo.github.io/PHASE-1-PROJECT-GERALD/

MIT License

Copyright (c) 2022 Kapgeradok

Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.